Fix: Prevent hardware watchdog in setup() function when extracting files
As writing to the filesystem is very slow, it needs a lot of time
extracting files from the image to the filesystem.
This caused to raise the hardware watchdog in setup() function.
To prevent raising the hardware watchdog, extraction of file content is
moved into loop() and writing is split into small blocks on each loop
step.